requirements.txt
streamlit
python-docx
openai
crewai
